requirements.txt
fastapi
uvicorn
python-multipart
deepface
psycopg2-binary
numpy
tf-keras
opencv-python-headless